+++ libs/mops/src/thds/mops/pure/__init__.py
@@ -21,6 +21,7 @@ from .core.source import create_source_at_uri  # noqa
 from .core.types import Args, BlobStore, Kwargs, Runner  # noqa
 from .core.uris import UriIsh, UriResolvable, register_blob_store  # noqa
 from .core.use_runner import use_runner  # noqa
+from .pickling import inject  # noqa
 from .pickling.memoize_only import memoize_in  # noqa
 from .pickling.mprunner import MemoizingPicklingRunner  # noqa
 from .runner.simple_shims import samethread_shim, subprocess_shim  # noqa